Refactor: A arquitetura do software foi alterada, ficando mais componentizada e utilizando context
