+++ frontend/src/context/ToDoContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const ToDoContext = createContext();
+
+export default ToDoContext;
